Continuation: Fixes #46 : Uploadable Features (#55)
* Update main Readme

* Update general docs

* Initial work outputting media objects LD formatted with IRI/Schema and embedded context

* Initial work outputting media objects LD formatted with IRI/Schema and embedded context

* Corrected expected json schema and output dummy to make post tests pass. Refactoring required to make the output MediaObject work properly now

* Uploadable docs update

* Output Correct primary MediaObject for the configured UploadableField annotated properties

* Support DataURI files as encoded by symfony normalizer. Additional support for url encoded text files

* Test encoding of base64 encoded without data: etc.

* Dynamically define filesystem services and update docs for Imagine

Dynamically creating the filesystems as services in a compiler pass.
Documented how we intend to support Imagine Bundle. Integration not yet complete.

* Fix imagine service wiring with factory for filesystem services

* Finish docs and methods to configure flyssytem into imagine

* Comment in services definition the properties that are injected at runtime

* Tests for imagine media objects - mocked out success response. Refactor to return real values

* Events added with dummy listeners for imagine cache store. Need to fix data loader to load source/uploaded file

* Cofigured data loader to use adapter set prior to being called from uploadablehelper. Events triggered in imagine cache lifecycle. Data storinggit add . Need to save file binary metadata to database still (field mapped).

* Fix entity mapping and convert to XML

* Saving imagine cache and returning media objects with correct values. Todo: endpoint to get the file stored primarily by the API. Possible proxy to imagine cache for private storage

* Fix symfony insights errors and configure to use php 7.4

* Further Symfony Insight Fixes

+++ src/Uploadable/UploadableHelper.php
@@ -13,14 +13,22 @@ declare(strict_types=1);
 
 namespace Silverback\ApiComponentsBundle\Uploadable;
 
+use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Persistence\ManagerRegistry;
-use Ramsey\Uuid\Uuid;
+use League\Flysystem\Filesystem;
+use League\Flysystem\UnableToReadFile;
+use Liip\ImagineBundle\Service\FilterService;
 use Silverback\ApiComponentsBundle\Annotation\UploadableField;
 use Silverback\ApiComponentsBundle\AnnotationReader\UploadableAnnotationReader;
+use Silverback\ApiComponentsBundle\Entity\Utility\ImagineFiltersInterface;
+use Silverback\ApiComponentsBundle\Factory\Uploadable\MediaObjectFactory;
 use Silverback\ApiComponentsBundle\Flysystem\FilesystemProvider;
-use Silverback\ApiComponentsBundle\Model\Uploadable\UploadedBase64EncodedFile;
+use Silverback\ApiComponentsBundle\Imagine\FlysystemDataLoader;
+use Silverback\ApiComponentsBundle\Model\Uploadable\MediaObject;
+use Silverback\ApiComponentsBundle\Model\Uploadable\UploadedDataUriFile;
 use Silverback\ApiComponentsBundle\Utility\ClassMetadataTrait;
 use Symfony\Component\HttpFoundation\FileBag;
+use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\PropertyAccess\PropertyAccess;
 
 /**
@@ -32,15 +40,27 @@ class UploadableHelper
 
     private UploadableAnnotationReader $annotationReader;
     private FilesystemProvider $filesystemProvider;
+    private MediaObjectFactory $mediaObjectFactory;
+    private RequestStack $requestStack;
+    private ?FilterService $filterService;
+    private FlysystemDataLoader $flysystemDataLoader;
 
     public function __construct(
         ManagerRegistry $registry,
         UploadableAnnotationReader $annotationReader,
-        FilesystemProvider $filesystemProvider
+        FilesystemProvider $filesystemProvider,
+        MediaObjectFactory $mediaObjectFactory,
+        RequestStack $requestStack,
+        FlysystemDataLoader $flysystemDataLoader,
+        ?FilterService $filterService = null
     ) {
         $this->initRegistry($registry);
         $this->annotationReader = $annotationReader;
         $this->filesystemProvider = $filesystemProvider;
+        $this->mediaObjectFactory = $mediaObjectFactory;
+        $this->requestStack = $requestStack;
+        $this->flysystemDataLoader = $flysystemDataLoader;
+        $this->filterService = $filterService;
     }
 
     public function setUploadedFilesFromFileBag(object $object, FileBag $fileBag): void
@@ -58,12 +78,32 @@ class UploadableHelper
         }
     }
 
+    public function storeFilesMetadata(object $object): void
+    {
+        $configuredProperties = $this->annotationReader->getConfiguredProperties($object, true, true);
+        $classMetadata = $this->getClassMetadata($object);
+
+        foreach ($configuredProperties as $fileProperty => $fieldConfiguration) {
+            // Let the data loader which should be configured for imagine to know which adapter to use
+            $this->flysystemDataLoader->setAdapter($fieldConfiguration->adapter);
+
+            $filename = $classMetadata->getFieldValue($object, $fieldConfiguration->property);
+            if ($object instanceof ImagineFiltersInterface && $this->filterService) {
+                $filters = $object->getImagineFilters(null);
+                foreach ($filters as $filter) {
+                    // This will trigger the cached file to be store
+                    // When cached files are store we save the file info
+                    $this->filterService->getUrlOfFilteredImage($filename, $filter);
+                }
+            }
+        }
+    }
+
     public function persistFiles(object $object): void
     {
         $propertyAccessor = PropertyAccess::createPropertyAccessor();
         $classMetadata = $this->getClassMetadata($object);
 
-        // $configuration = $this->annotationReader->getConfiguration($object);
         $configuredProperties = $this->annotationReader->getConfiguredProperties($object, true, true);
         /**
          * @var UploadableField[] $configuredProperties
@@ -73,17 +113,17 @@ class UploadableHelper
             if ($currentFilepath) {
                 $this->removeFilepath($object, $fieldConfiguration);
             }
-            /** @var UploadedBase64EncodedFile|null $file */
+            /** @var UploadedDataUriFile|null $file */
             $file = $propertyAccessor->getValue($object, $fileProperty);
             if (!$file) {
                 $classMetadata->setFieldValue($object, $fieldConfiguration->property, null);
                 continue;
             }
 
-            $filesystem = $this->filesystemProvider->getFilesystem($fieldConfiguration->adapter);
+            $filesystem = $this->getFilesystemFromFieldConfiguration($fieldConfiguration);
 
             $path = $fieldConfiguration->prefix ?? '';
-            $path .= Uuid::uuid4() . $file->getFilename();
+            $path .= $file->getFilename();
             $stream = fopen($file->getRealPath(), 'r');
             $filesystem->writeStream($path, $stream, [
                 'mimetype' => $file->getMimeType(),
@@ -93,12 +133,88 @@ class UploadableHelper
         }
     }
 
+    public function deleteFiles(object $object): void
+    {
+        $classMetadata = $this->getClassMetadata($object);
+
+        $configuredProperties = $this->annotationReader->getConfiguredProperties($object, true, true);
+        foreach ($configuredProperties as $fileProperty => $fieldConfiguration) {
+            $currentFilepath = $classMetadata->getFieldValue($object, $fieldConfiguration->property);
+            if ($currentFilepath) {
+                $this->removeFilepath($object, $fieldConfiguration);
+            }
+        }
+    }
+
+    public function getMediaObjects(object $object): ?ArrayCollection
+    {
+        $collection = new ArrayCollection();
+        $classMetadata = $this->getClassMetadata($object);
+
+        $configuredProperties = $this->annotationReader->getConfiguredProperties($object, true, true);
+        foreach ($configuredProperties as $fileProperty => $fieldConfiguration) {
+            $propertyMediaObjects = [];
+            $filesystem = $this->getFilesystemFromFieldConfiguration($fieldConfiguration);
+            $path = $classMetadata->getFieldValue($object, $fieldConfiguration->property);
+            if (!$path) {
+                continue;
+            }
+            if (!$filesystem->fileExists($path)) {
+                continue;
+            }
+
+            // Populate the primary MediaObject
+            try {
+                $propertyMediaObjects[] = $this->mediaObjectFactory->create($filesystem, $path);
+            } catch (UnableToReadFile $exception) {
+            }
+
+            if ($object instanceof ImagineFiltersInterface) {
+                array_push($propertyMediaObjects, ...$this->getMediaObjectsForImagineFilters($object, $path, $fieldConfiguration->adapter));
+            }
+
+            $collection->set($fieldConfiguration->property, $propertyMediaObjects);
+        }
+
+        return $collection->count() ? $collection : null;
+    }
+
+    /**
+     * @return MediaObject[]
+     */
+    private function getMediaObjectsForImagineFilters(ImagineFiltersInterface $object, string $path, string $adapter): array
+    {
+        // Let the data loader which should be configured for imagine to know which adapter to use
+        $this->flysystemDataLoader->setAdapter($adapter);
+
+        $mediaObjects = [];
+        if (!$this->filterService) {
+            return $mediaObjects;
+        }
+
+        $request = $this->requestStack->getMasterRequest();
+        $filters = $object->getImagineFilters($request);
+        foreach ($filters as $filter) {
+            $resolvedUrl = $this->filterService->getUrlOfFilteredImage($path, $filter);
+            $mediaObjects[] = $this->mediaObjectFactory->createFromImagine($resolvedUrl, $path, $filter);
+        }
+
+        return $mediaObjects;
+    }
+
+    private function getFilesystemFromFieldConfiguration(UploadableField $fieldConfiguration): Filesystem
+    {
+        return $this->filesystemProvider->getFilesystem($fieldConfiguration->adapter);
+    }
+
     private function removeFilepath(object $object, UploadableField $fieldConfiguration): void
     {
         $classMetadata = $this->getClassMetadata($object);
 
         $filesystem = $this->filesystemProvider->getFilesystem($fieldConfiguration->adapter);
         $currentFilepath = $classMetadata->getFieldValue($object, $fieldConfiguration->property);
-        $filesystem->delete($currentFilepath);
+        if ($filesystem->fileExists($currentFilepath)) {
+            $filesystem->delete($currentFilepath);
+        }
     }
 }
